// src/types/typesApi.ts
export type typesApi = {
    address_1: string;
    brewery_type: string;
    state: string;
    city: string;
    country: string;
    name: string;
    postal_code: string;
    street: string;
    id: string;
    website_url: string;
    longitude: string;
    latitude: string;
    phone: string;
}